poprawa błędów

+++ DatabaseConnectionDll/DatabaseConnection/Repositories/IRepository.cs
@@ -5,7 +5,7 @@ namespace DatabaseConnection.Repositories
     public interface IRepository<T> where T : class
     {
         IList<T> GetAll();
-        T Get(long id);
+        T Get(int id);
         void Add(T entity);
         void AddRange(List<T> entities);
         void Delete(T entity);
